WIP Global Configuration interface

+++ src/config.ts
@@ -1,9 +1,10 @@
 import {Schema, Document} from 'mongoose';
 import {KoratCore} from './core';
+import {KoratValue, KoratValueConstraints, createValue} from './data';
 
 export interface KoratConfig {
-  getEntry: (key: string) => void,
-  addEntry: (key: string, entry?: KoratConfigEntry) => void,
+  getEntry: (key: string) => KoratValue,
+  addEntry: (key: string, entry?: KoratValue) => void,
   entryExists: (key: string) => boolean,
   getValue: (key: string) => any,
   setValue: (key: string, value: any) => void,
@@ -11,38 +12,40 @@ export interface KoratConfig {
   loadEntries: () => Promise<void>,
 }
 
-export interface KoratConfigEntry {
-  value?: any,
-}
-
 interface ConfigDocument extends Document {
   key: string,
-  value: any
+  value: any,
+  constraints: object,
 }
 
 const configSchema = new Schema({
   key: {type: String, required: true, unique: true},
-  value: Object,
+  value: {type: Object},
+  constraints: {type: Object},
 });
 
 export function createConfig(core: KoratCore): KoratConfig {
-  const {databaseManager, events} = core;
+  const {server} = core;
 
-  databaseManager.createModel('Config', configSchema);
-  events.listen('server-started', () => core.config.loadEntries());
-  events.listen('server-stopping', () => core.config.saveEntries());
+  server.afterStart(async () => {
+    server.mongooseConnection && server.mongooseConnection.model('Config', configSchema);
+    await core.config.loadEntries();
+  });
+  server.beforeStop(async () => {
+    await core.config.saveEntries();
+  });
 
-  const entries: Record<string, KoratConfigEntry> = {};
+  const entries: Record<string, KoratValue> = {};
 
   return {
     getEntry(key) {
       return entries[key];
     },
 
-    addEntry(key, entry = {}) {
+    addEntry(key, value) {
       if (this.entryExists(key))
         throw new Error(`Config entry with key ${key} already exists.`);
-      entries[key] = entry || {};
+      entries[key] = value || createValue();
     },
 
     entryExists(key) {
@@ -60,22 +63,25 @@ export function createConfig(core: KoratCore): KoratConfig {
     },
 
     async saveEntries() {
-      const ConfigModel = databaseManager.getModel<ConfigDocument>('Config');
+      const entriesToSave = Object.entries(entries).map(([key, entry]) => ({
+        key,
+        value: entry.value,
+        constraints: entry.constraints
+      }));
+
+      const ConfigModel = server.mongooseConnection && server.mongooseConnection.model<ConfigDocument>('Config');
+      if (!ConfigModel) throw new Error();
 
       await ConfigModel.deleteMany({});
-      await ConfigModel.insertMany(Object.keys(entries).map(key => ({
-        key,
-        value: this.getValue(key)
-      })));
+      await ConfigModel.insertMany(entriesToSave);
     },
 
     async loadEntries() {
-      const ConfigModel = databaseManager.getModel<ConfigDocument>('Config');
+      const ConfigModel = server.mongooseConnection && server.mongooseConnection.model<ConfigDocument>('Config');
+      if (!ConfigModel) throw new Error();
 
       for (let entry of await ConfigModel.find()) {
-        this.addEntry(entry.key, {
-          value: entry.value
-        })
+        this.addEntry(entry.key, createValue(entry.value, <KoratValueConstraints>entry.constraints));
       }
     }
   };
